feat: separate profile card from view
